meiduo_mall/requirements.txt
amqp==2.1.4
anyjson==0.3.3
appdirs==1.4.3
billiard==3.5.0.2
celery==4.0.2
certifi==2018.10.15
chardet==3.0.4
decorator==4.0.11
Django==1.11.11
django-ckeditor==5.6.1
django-cors-headers==2.4.0
django-crontab==0.7.1
django-haystack==2.8.1
django-js-asset==1.1.0
django-redis==4.8.0
django-redis-sessions==0.5.6
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
drf-extensions==0.4.0
drf-haystack==1.8.4
elasticsearch==2.4.1
fdfs-client-py==1.2.6
idna==2.7
ipdb==0.10.3
ipython==6.0.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.10.2
jieba==0.38
jsonfield==2.0.1
kombu==4.0.2
mutagen==1.41.1
olefile==0.44
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
pkg-resources==0.0.0
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
Pygments==2.2.0
PyJWT==1.6.4
pymongo==3.4.0
PyMySQL==0.7.10
pyparsing==2.2.0
python-dateutil==2.7.5
pytz==2016.10
redis==2.10.5
requests==2.20.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
urllib3==1.24.1
vine==1.1.3
wcwidth==0.1.7
Whoosh==2.7.4
